Add fhost queue squad btn

+++ events/buttons/fhost.js
@@ -27,11 +27,24 @@ module.exports = {
         })
         let hostComps = i.message.components.map(x => x.components).flat();
         const allIDs = hostFields.map(x => x[1].split('|')).flat()
-        if (allIDs.includes(i.user.id) && i.customId != 'fhost-❌') return i.update({ });
+        let minimumFill = 4;
+        let doFilterCheck = true;
 
-        const newRow = (arr) => new ActionRowBuilder().addComponents(arr)
+        if (i.customId === 'fhost-✅' && i.message.content.slice(2, -1) == i.user.id) {
+            const filterCheck = allIDs.filter(x => x != 'n');
+            if (filterCheck.length <= 1) {
+                return i.reply({ content: `Squad needs atleast 2 members.`, ephemeral: true })
+            } else {
+                minimumFill = filterCheck;
+                doFilterCheck = false;
+            }
+        }
+
+        if (doFilterCheck && allIDs.includes(i.user.id) && i.customId != 'fhost-❌') return i.update({ });
         
-        if (i.customId == 'fhost-❌' && allIDs.includes(i.user.id) && i.message.content.slice(2, -1) != i.user.id) {
+        const newRow = (arr) => new ActionRowBuilder().addComponents(arr);
+
+        if (doFilterCheck && i.customId == 'fhost-❌' && allIDs.includes(i.user.id) && i.message.content.slice(2, -1) != i.user.id) {
             const joinedField = hostFields.findIndex(x => x[1].split('|').includes(i.user.id))
             const joinedButton = hostComps.findIndex(x => x.customId == `fhost-${hostFields[joinedField][0]}`)
             if (hostComps[joinedButton].customId != 'fhost-Any') {
@@ -43,30 +56,35 @@ module.exports = {
             globalHostEmbed.addFields(...hostFields.map(x => { return { name: x[0], value: x[1], inline: true } })).setTimestamp()
             allIDs[joinedButton] = 'n'
             await i.update({ content: i.message.content, embeds: [globalHostEmbed], components: [newRow(hostComps.slice(0, 5)), newRow(hostComps.slice(5))] })
-        } else if (i.customId == 'fhost-❌' && allIDs.includes(i.user.id) && i.message.content.slice(2, -1) == i.user.id) {
+        } else if (doFilterCheck && i.customId == 'fhost-❌' && allIDs.includes(i.user.id) && i.message.content.slice(2, -1) == i.user.id) {
             await i.message.delete();
             await i.channel.send({ embeds: [
                 new EmbedBuilder().setTitle(`Run "${hostTitle}" was cancelled`)
             ] })
         } else {
-            const getField = hostFields.findIndex(x => x[0] == i.customId.split('-')[1]);
-            if (getField === -1) return i.update({ });
-            if (hostFields[getField][1] != 'n' && i.customId != 'fhost-Any') return await i.update({ });
-            let currentField = [...hostFields[getField][1].split('|')].filter(x => x != 'n')
-            currentField.push(i.user.id)
-            hostFields[getField][1] = currentField.join('|')
-            
-            const clickedButton = hostComps.findIndex(x => x.customId == `fhost-${hostFields[getField][0]}`)
-            if (hostComps[clickedButton].customId !== `fhost-Any`) {
-                hostComps[clickedButton].data.disabled = !hostComps[clickedButton].disabled;
+            if (doFilterCheck) {
+                const getField = hostFields.findIndex(x => x[0] == i.customId.split('-')[1]);
+                if (getField === -1) return i.update({ });
+                if (hostFields[getField][1] != 'n' && i.customId != 'fhost-Any') return await i.update({ });
+                let currentField = [...hostFields[getField][1].split('|')].filter(x => x != 'n')
+                currentField.push(i.user.id)
+                hostFields[getField][1] = currentField.join('|')
+    
+                const clickedButton = hostComps.findIndex(x => x.customId == `fhost-${hostFields[getField][0]}`)
+                if (hostComps[clickedButton].customId !== `fhost-Any`) {
+                    hostComps[clickedButton].data.disabled = !hostComps[clickedButton].disabled;
+                }
+    
+                hostFields = hostFields.map(x => [x[0], x[1].split('|').map(y => y == 'n' || !y ? 'None' : `<@${y}>`).join(', ')])
+                globalHostEmbed.addFields(...hostFields.map(x => { return { name: x[0], value: x[1], inline: true } })).setTimestamp()
+                allIDs.push(i.user.id)
+            } else {
+                globalHostEmbed = i.message.embeds[0]
             }
-            hostFields = hostFields.map(x => [x[0], x[1].split('|').map(y => y == 'n' || !y ? 'None' : `<@${y}>`).join(', ')])
-            globalHostEmbed.addFields(...hostFields.map(x => { return { name: x[0], value: x[1], inline: true } })).setTimestamp()
-            allIDs.push(i.user.id)
             await i.update({ content: i.message.content, embeds: [globalHostEmbed], components: [newRow(hostComps.slice(0, 5)), newRow(hostComps.slice(5))] })
         }
 
-        if (allIDs.filter(x => x != 'n').length >= 4) {
+        if (allIDs.filter(x => x != 'n').length >= minimumFill.length) {
             await i.message.edit({ content: 'Preparing squad...', components: [] });
             let names = await getAllUserData('farmer')
             await i.message.delete();
